Add a testing module for testing helper functions
