+++ signer/src/testing.rs
@@ -1,0 +1,40 @@
+//! Module with testing utility functions.
+
+use bitcoin::key::TapTweak;
+use bitcoin::TapSighashType;
+use bitcoin::Witness;
+use secp256k1::SECP256K1;
+use crate::utxo::UnsignedTransaction;
+
+
+/// A helper function for correctly setting witness data
+pub fn set_witness_data(unsigned: &mut UnsignedTransaction, keypair: secp256k1::Keypair) {
+    let sighash_type = TapSighashType::Default;
+    let sighashes = unsigned.construct_digests().unwrap();
+
+    let signer_msg = secp256k1::Message::from(sighashes.signers);
+    let tweaked = keypair.tap_tweak(SECP256K1, None);
+    let signature = SECP256K1.sign_schnorr(&signer_msg, &tweaked.to_inner());
+    let signature = bitcoin::taproot::Signature { signature, sighash_type };
+    let signer_witness = Witness::p2tr_key_spend(&signature);
+
+    let deposit_witness = sighashes.deposits.into_iter().map(|(deposit, sighash)| {
+        let deposit_msg = secp256k1::Message::from(sighash);
+        let signature = SECP256K1.sign_schnorr(&deposit_msg, &keypair);
+        let signature = bitcoin::taproot::Signature { signature, sighash_type };
+        deposit.construct_witness_data(signature)
+    });
+
+    let witness_data: Vec<Witness> = std::iter::once(signer_witness)
+        .chain(deposit_witness)
+        .collect();
+
+    unsigned
+        .tx
+        .input
+        .iter_mut()
+        .zip(witness_data)
+        .for_each(|(tx_in, witness)| {
+            tx_in.witness = witness;
+        });
+}
