+++ signer/tests/integration/regtest.rs
@@ -33,8 +33,9 @@ use bitcoincore_rpc::Client;
 use bitcoincore_rpc::Error as BtcRpcError;
 use bitcoincore_rpc::RpcApi;
 use secp256k1::SECP256K1;
-use signer::utxo::UnsignedTransaction;
 use std::sync::OnceLock;
+
+pub use signer::testing::set_witness_data;
 
 /// These must match the username and password in bitcoin.conf
 const BITCOIN_CORE_RPC_USERNAME: &str = "alice";
@@ -369,35 +370,4 @@ pub fn p2tr_sign_transaction<U>(
     let signature = bitcoin::taproot::Signature { signature, sighash_type };
 
     tx.input[input_index].witness = Witness::p2tr_key_spend(&signature);
-}
-
-pub fn set_witness_data(unsigned: &mut UnsignedTransaction, keypair: secp256k1::Keypair) {
-    let sighash_type = TapSighashType::Default;
-    let sighashes = unsigned.construct_digests().unwrap();
-
-    let signer_msg = secp256k1::Message::from(sighashes.signers);
-    let tweaked = keypair.tap_tweak(SECP256K1, None);
-    let signature = SECP256K1.sign_schnorr(&signer_msg, &tweaked.to_inner());
-    let signature = bitcoin::taproot::Signature { signature, sighash_type };
-    let signer_witness = Witness::p2tr_key_spend(&signature);
-
-    let deposit_witness = sighashes.deposits.into_iter().map(|(deposit, sighash)| {
-        let deposit_msg = secp256k1::Message::from(sighash);
-        let signature = SECP256K1.sign_schnorr(&deposit_msg, &keypair);
-        let signature = bitcoin::taproot::Signature { signature, sighash_type };
-        deposit.construct_witness_data(signature)
-    });
-
-    let witness_data: Vec<Witness> = std::iter::once(signer_witness)
-        .chain(deposit_witness)
-        .collect();
-
-    unsigned
-        .tx
-        .input
-        .iter_mut()
-        .zip(witness_data)
-        .for_each(|(tx_in, witness)| {
-            tx_in.witness = witness;
-        });
 }
